feat: Update display of CO2 Gadget version in web server
The code changes update the display of the CO2 Gadget version in the web server. The `fetchVersion()` function is refactored into `displayVersion()` in both `preferences.js` and `status.js` files. The version information is fetched from the server using the fetch API and then displayed on the web page. The version text is formatted as "CO2 Gadget: v{major}.{minor}.{revision}-{branch} (Flavour: {flavour})" in `preferences.js` and "CO2 Gadget: {versionText}" in `status.js`.